move nsf gui to its own file
